Getting resource analysis into SDK.
